python/requirements.txt
exa-py>=1.0.0
voyageai>=0.3.0
python-dotenv>=1.0.0
supabase>=2.8.1
